Move language.ts and dependencies
