+++ packages/docs/src/components/header/components/LanguageDropdown.vue
@@ -33,7 +33,7 @@ import { computed, defineComponent } from 'vue'
 import { getColors } from 'vuestic-ui/src/main'
 import { useI18n } from 'vue-i18n'
 import { useRouter } from 'vue-router'
-import { languages } from './../../languages'
+import { languages } from '@/locales'
 
 export default defineComponent({
   setup () {
